requirements.txt
pyserial
flask